--- src/utils/error.ts
@@ -1,8 +1,0 @@
-export class HttpError extends Error {
-	statusCode: number;
-
-	constructor(message: string, statusCode: number) {
-		super(message);
-		this.statusCode = statusCode;
-	}
-}
